posttraining-research-ci-supp/components/privacylens/requirements.txt
streamlit
numpy
scipy
rouge_score
fire
transformers>=4.26.1
torch
vllm
openai==0.28.0
langchain==0.0.277
black
python-dotenv==1.0.0
tiktoken
wikipedia
chromadb
scipy
scikit-learn
statsmodels
matplotlib
seaborn
anthropic==0.3.6
accelerate
bitsandbytes
sentencepiece
pydantic
azure
git+https://github.com/dhh1995/PromptCoder.git#egg=procoder